+++ NEWS.md
@@ -3,6 +3,11 @@ Phan NEWS
 ??? ?? 2020, Phan 3.1.2 (dev)
 -----------------------
 
+New features (CLI, Config):
++ Add `--always-exit-successfully-after-analysis` flag.
+  By default, phan exits with a non-zero exit code if 1 or more unsuppressed issues were reported.
+  When this CLI flag is set, phan will instead exit with exit code 0 as long as the analysis completed.
+
 New features (Analysis):
 + Emit `PhanNoopRepeatedSilenceOperator` for `@@expr` or `@(@expr)`.
   This is less efficient and only makes a difference in extremely rare edge cases.